+++ rust/lance/src/io/commit.rs
@@ -53,7 +53,7 @@ use crate::dataset::fragment::FileFragment;
 use crate::dataset::transaction::{Operation, Transaction};
 use crate::dataset::{write_manifest_file, ManifestWriteConfig};
 use crate::format::{DeletionFile, Fragment};
-use crate::index::DatasetIndexInternalExt;
+use crate::index::{DatasetIndexExt, DatasetIndexInternalExt};
 use crate::Dataset;
 
 #[cfg(all(target_feature = "dynamodb", test))]
